backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.2
requests==2.32.3
neo4j==5.25.0
